#18230  Extracting core char to byte  conversion methods to a separate class
