+++ dotCMS/src/integration-test/java/com/dotcms/security/apps/AppsAPIImplTest.java
@@ -20,26 +20,17 @@ import com.dotmarketing.business.Role;
 import com.dotmarketing.business.portal.PortletAPI;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotSecurityException;
-import com.dotmarketing.util.Logger;
 import com.google.common.collect.ImmutableSet;
 import com.liferay.portal.model.Portlet;
 import com.liferay.portal.model.User;
-import com.tngtech.java.junit.dataprovider.DataProvider;
-import com.tngtech.java.junit.dataprovider.DataProviderRunner;
-import com.tngtech.java.junit.dataprovider.UseDataProvider;
 import io.vavr.Tuple2;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
 import org.apache.commons.lang.RandomStringUtils;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-@RunWith(DataProviderRunner.class)
 public class AppsAPIImplTest {
 
     @BeforeClass
@@ -339,84 +330,6 @@ public class AppsAPIImplTest {
         api.saveSecrets(builder.build(), APILocator.systemHost() , nonAdminUserWithAccessToPortlet);
     }
 
-    @DataProvider
-    public static Object[] getTestCases() throws Exception {
-        return new Object[]{
-                new UTFCharsRangeTestCase(0, 127, "C0 Controls and Basic Latin"),
-                new UTFCharsRangeTestCase(128, 255, "C1 Controls and Latin-1 Supplement"),
-                new UTFCharsRangeTestCase(256, 383, "Latin Extended-A"),
-                new UTFCharsRangeTestCase(384, 591, "Latin Extended-B"),
-                new UTFCharsRangeTestCase(688, 767, "Spacing Modifiers"),
-                new UTFCharsRangeTestCase(768, 879, "Diacritical Marks"),
-                new UTFCharsRangeTestCase(880, 1023, "Greek and Coptic"),
-                new UTFCharsRangeTestCase(1024, 1279, "Cyrillic Basic"),
-                new UTFCharsRangeTestCase(1280, 1327, "Cyrillic Supplement"),
-                new UTFCharsRangeTestCase(8192, 8303, "General Punctuation"),
-                new UTFCharsRangeTestCase(8352, 8399, "Currency Symbols"),
-                new UTFCharsRangeTestCase(8448, 8527, "Letterlike Symbols"),
-                new UTFCharsRangeTestCase(8592, 8703, "Arrows"),
-                new UTFCharsRangeTestCase(8704, 8959, "Mathematical Operators"),
-                new UTFCharsRangeTestCase(9472, 9599, "Box Drawings"),
-                new UTFCharsRangeTestCase(9600, 9631, "Block Elements"),
-                new UTFCharsRangeTestCase(9632, 9727, "Geometric Shapes"),
-                new UTFCharsRangeTestCase(9728, 9983, "Miscellaneous Symbols"),
-                new UTFCharsRangeTestCase(9984, 10175, "Dingbats")
-        };
-    }
-
-    
-
-    @Test
-    @UseDataProvider("getTestCases")
-    public void Test_BytesToChars_No_Middle_String_Conversion(final UTFCharsRangeTestCase testCase) throws IOException {
-        final AppsAPIImpl impl = new AppsAPIImpl();
-        final StringBuilder stringBuilder = new StringBuilder();
-        for(int i=  testCase.fromCode; i<= testCase.toCode; i++) {
-            final String string = fromCharCode(i);
-            stringBuilder.append(string);
-        }
-        final String input = stringBuilder.toString();
-        Logger.info(AppsAPIImplTest.class,()->  String.format(" UTF Charset code from `%d` to `%d`  %s `%s` ",testCase.fromCode, testCase.toCode, testCase.description,input));
-        final char [] chars = impl.bytesToCharArrayUTF(input.getBytes(StandardCharsets.UTF_8));
-        final byte [] bytes = impl.charsToBytesUTF(chars);
-        final String output = new String(bytes, StandardCharsets.UTF_8);
-        assertEquals(input, output);
-    }
-
-    /**
-     * https://www.w3schools.com/charsets/ref_html_utf8.asp
-     * @param codePoints char code see utf char codes.
-     * @return the utf string representation.
-     */
-    private static String fromCharCode(final int... codePoints) {
-        return new String(codePoints, 0, codePoints.length);
-    }
-
-    @Test
-    public void Test_Secret_Json_Serialization_No_String_Middle_Man() throws DotDataException {
-        final AppsAPIImpl impl = new AppsAPIImpl();
-        final AppSecrets secretsIn = new AppSecrets.Builder()
-                .withKey("TheKey")
-                .withHiddenSecret("hidden1", "I'm hidden")
-                .withSecret("non-hidden1", "I'm not hidden")
-                .withSecret("non-hidden5", RandomStringUtils.randomAlphanumeric(2337))
-                .withSecret("bool1", true)
-                .build();
-        final char[] toJsonAsChars = impl.toJsonAsChars(secretsIn);
-        final AppSecrets secretsOut = impl.readJson(toJsonAsChars);
-        assertEquals(secretsIn.getKey(), secretsOut.getKey());
-
-        assertEquals(secretsIn.getSecrets().size(), secretsOut.getSecrets().size());
-
-        final Set<Entry<String, Secret>> secretsInEntries = secretsIn.getSecrets().entrySet();
-        for (final Entry<String, Secret> entryIn : secretsInEntries) {
-            assertEquals(secretsIn.getKey(), secretsOut.getKey());
-            final Secret out = secretsOut.getSecrets().get(entryIn.getKey());
-            assertNotNull(out);
-            assertTrue(out.equals(entryIn.getValue()));//This does a deepEquals.
-        }
-    }
-
     @Test
     public void Test_Secret_Destroy_Method() {
         final AppSecrets secrets = new AppSecrets.Builder()
@@ -443,19 +356,6 @@ public class AppsAPIImplTest {
             portlet = portletDataGen.portletId(integrationsPortletId).nextPersisted();
         }
         return portlet;
-    }
-
-    static class UTFCharsRangeTestCase {
-
-        final int fromCode;
-        final int toCode;
-        final String description;
-
-        UTFCharsRangeTestCase(final int fromCode, final int toCode, final String description) {
-            this.fromCode = fromCode;
-            this.toCode = toCode;
-            this.description = description;
-        }
     }
 
 }
